+++ modules/log/index.js
@@ -17,6 +17,49 @@ module.exports.readConfig = function (configIn) {
 	}
 	
 	config = configIn;
+
+	if (config.level && typeof config.level === 'string') {
+		// we now support a string format of level
+		// e.i. "level": ">= info"
+		var sep = config.level.split(' ');
+		var operators = ['>', '<', '>=', '<=', '='];
+		var levels = ['verbose', 'debug', 'info', 'warning', 'error', 'fatal'];
+		var level = {};
+		var op = null;
+		var lvl = null;
+		for (var k = 0, ken = sep.length; k < ken; k++) {
+			if (operators.indexOf(sep[k]) !== -1) {
+				op = sep[k];
+			} else if (levels.indexOf(sep[k]) !== -1) {
+				lvl = sep[k];
+			}
+		}
+		if (lvl) {
+			if (op) {
+				if (op.indexOf('<') !== -1) {
+					levels.reverse();
+				}
+				var start = levels.indexOf(lvl);
+				if (op.indexOf('=') === -1) {
+					start += 1;
+				}
+				for (var j = start, jen = levels.length; j < jen; j++) {
+					level[levels[j]] = true;
+				}
+			} else {
+				level[lvl] = true;
+			}
+		}
+		config.level = level;
+
+	} else if (config.level && Array.isArray(config.level)) {
+		// we now support an array format of level
+		var levelObj = {};
+		for (var i = 0, len = config.level.length; i < len; i++) {
+			levelObj[config.level[i]] = true;
+		}
+		config.level = levelObj;
+	}
 	
 	return true;
 };